+++ src/decorators/require-decorators.ts
@@ -15,19 +15,64 @@ import * as ReflectMetadata from 'reflect-metadata';
 
 import { EgeoUtils } from '../utils';
 
-const REQUIRED_METADATA = Symbol('EgeoRequire');
+const EGEO_METADATA = Symbol('EgeoLibraryMetadata');
 
 // tslint:disable:only-arrow-functions
 export function Required(condition?: string): any {
    return function (target: any, name: string): any {
-      const meta: any = Reflect.getOwnMetadata(REQUIRED_METADATA, target.constructor) || {};
+      const meta: any = Reflect.getOwnMetadata(EGEO_METADATA, target.constructor) || {};
       meta[name] = meta.hasOwnProperty(name) && meta[name] || { type: METADATA_TYPE.PROPERTY };
       meta[name].required = true;
       if (condition !== undefined) {
          meta[name].requireCondition = condition;
       }
 
-      Reflect.defineMetadata(REQUIRED_METADATA, meta, target.constructor);
+      Reflect.defineMetadata(EGEO_METADATA, meta, target.constructor);
+   };
+}
+
+// tslint:disable:only-arrow-functions
+export function EgRequired(condition?: string): any {
+   return function (target: any, name: string): any {
+      const meta: any = Reflect.getOwnMetadata(EGEO_METADATA, target.constructor) || {};
+      meta[name] = meta.hasOwnProperty(name) && meta[name] || { type: METADATA_TYPE.PROPERTY };
+      meta[name].required = true;
+      if (condition !== undefined) {
+         meta[name].requireCondition = condition;
+      }
+
+      Reflect.defineMetadata(EGEO_METADATA, meta, target.constructor);
+   };
+}
+
+export function EgDeprecated(initialValue?: any): any {
+   return function (target: any, name: string): any {
+      const meta: any = Reflect.getOwnMetadata(EGEO_METADATA, target.constructor) || {};
+      meta[name] = meta.hasOwnProperty(name) && meta[name] || { type: METADATA_TYPE.PROPERTY };
+      meta[name].deprecated = true;
+      if (initialValue !== undefined) {
+         meta[name].initialValue = initialValue;
+      }
+
+      Reflect.defineMetadata(EGEO_METADATA, meta, target.constructor);
+   };
+}
+
+export function Egeo(params?: ''): any {
+   return function (target: any): any {
+      let _onInit = target.prototype.ngOnInit;
+      if (_onInit !== undefined) {
+         target.prototype.ngOnInit = function (...args: any[]): void {
+            checkDeprecated(target, this);
+            checkRequired(target, this);
+            _onInit.apply(this, args);
+         };
+      } else {
+         target.prototype.ngOnInit = function (): void {
+            checkDeprecated(target, this);
+            checkRequired(target, this);
+         };
+      }
    };
 }
 
@@ -47,8 +92,23 @@ export function CheckRequired(params?: ''): any {
    };
 }
 
+function checkDeprecated(target: any, scope: any): void {
+   const meta: any = Reflect.getOwnMetadata(EGEO_METADATA, target);
+   if (meta !== undefined) {
+      Object.keys(meta).forEach((key) => {
+         if (meta[key].deprecated) {
+            if (scope[key] !== undefined) {
+               console.warn(`${EgeoUtils.toDash(target.name)}: field ${key} is a deprecated field`);
+            } else if (meta[key].initialValue) {
+               scope[key] = meta[key].initialValue;
+            }
+         }
+      });
+   }
+}
+
 function checkRequired(target: any, scope: any): void {
-   const meta: any = Reflect.getOwnMetadata(REQUIRED_METADATA, target);
+   const meta: any = Reflect.getOwnMetadata(EGEO_METADATA, target);
    if (meta !== undefined) {
       let inputs: string[] = getKeys(Object.keys(meta), meta, scope);
       EgeoUtils.validateInputs(scope, inputs, target.name);
@@ -82,5 +142,7 @@ export interface EgeoMetadata {
    type: METADATA_TYPE;
    required?: boolean;
    requireCondition?: string;
+   deprecated?: boolean;
+   initialValue?: any;
 }
 // tslint:enabled